+++ db/math_review_list.sql
@@ -1,5 +1,5 @@
 CREATE TABLE /*_*/math_review_list (
-  revision_id INT(11)     NOT NULL,
+  revision_id BIGINT(20)     NOT NULL,
   anchor      VARCHAR(50) NOT NULL,
   priority    TINYINT(4)  NOT NULL,
   UNIQUE KEY ix_math_rev_list (revision_id, anchor)
