+++ db/mathsemantics.sql
@@ -1,5 +1,5 @@
 CREATE TABLE /*_*/mathsemantics (
-  `revision_id` int(10) UNSIGNED NOT NULL,
+  `revision_id` bigint(20) UNSIGNED NOT NULL,
   `identifier` varchar(20) NOT NULL,
   `evidence` double NOT NULL,
   `noun` varchar(255) NOT NULL,
