+++ db/wmc/persistent/math_wmc_ref.sql
@@ -1,6 +1,6 @@
 CREATE TABLE math_wmc_ref (
     qId INT PRIMARY KEY NOT NULL,
-    oldId INT,
+    oldId BIGINT,
     fId INT,
     qVarCount SMALLINT,
     texQuery TEXT,
